Fixed addTwoInts service bug. (Responce msg should contain the size of the msg after the 'ok')

+++ src/utils.c
@@ -2,9 +2,9 @@
 
 #include <unistd.h>
 
-int utils_get_int_from_bytes(const char* bytes, int nbytes)
+int32_t utils_get_int_from_bytes(const char* bytes, int nbytes)
 {
-	int res = 0, i;
+	int32_t res = 0, i;
 	if(nbytes == 4)
 	{
 		res = (bytes[3] << 24) | (bytes[2] << 16) | (bytes[1] << 8) | bytes[0];
@@ -19,31 +19,31 @@ int utils_get_int_from_bytes(const char* bytes, int nbytes)
 	return res;
 }
 
-long utils_get_long_from_bytes(const char* bytes, int nbytes)
+int64_t utils_get_long_from_bytes(const char* bytes, int nbytes)
 {
-	long res = 0l;
+	int64_t res = 0l;
 	int i;
 
 	if(nbytes == 8)
 	{
-		res = ((long)(0xff & bytes[7]) << 56) |
-		 	  ((long)(0xff & bytes[6]) << 48) |
-		 	  ((long)(0xff & bytes[5]) << 40) |
-		 	  ((long)(0xff & bytes[4]) << 32) |
-			  ((long)(0xff & bytes[3]) << 24) |
-		 	  ((long)(0xff & bytes[2]) << 16) |
-		 	  ((long)(0xff & bytes[1]) << 8) |
-		 	  (long)(0xff & bytes[0]);
+		res = ((int64_t)(0xff & bytes[7]) << 56) |
+		 	  ((int64_t)(0xff & bytes[6]) << 48) |
+		 	  ((int64_t)(0xff & bytes[5]) << 40) |
+		 	  ((int64_t)(0xff & bytes[4]) << 32) |
+			  ((int64_t)(0xff & bytes[3]) << 24) |
+		 	  ((int64_t)(0xff & bytes[2]) << 16) |
+		 	  ((int64_t)(0xff & bytes[1]) << 8) |
+		 	   (int64_t)(0xff & bytes[0]);
 	}
 	else
 	{
 		for(i = 0; i < nbytes; ++i)
 		{
-			res |= (long)(0xff & bytes[i]) << (i * 8);
+			res |= (int64_t)(0xff & bytes[i]) << (i * 8);
 		}
 	}
 
-	return (long)res;
+	return res;
 }
 
 int utils_get_pid()
